webgpu: refactor atomicAdd code
